Change how blocks are drawn; class change
Blocks are now drawn proportionally to the x axis.
This means blocks will appear symmetric to the player from both ends.

This is a desire I wanted to see in the program,
as it makes the game modifiable and more pleasant to the user's eyes.

The block class was also changed.
Rects per column/row have increased.

The block class has changed because the
original algorithm did not layer Rects correctly.