Change how enemies move in block
Enemies will now move on both ends of the screen as a block.

Once an enemy reaches the end of the screen (x axis),
the entire block will move down
and in the opposite direction.

Fixed bug where destroying an enemy would require the user to
press the space-bar twice to fire again.