Abstract baseobjects; Change block STRUCTURE
Baseobjects are now known as gameobjects.
A lot of the smaller gameobject classes were moved into the
same module as gameobject.

Gameobjects to me just made more sense than 'baseobjects'.
A lot of the classes were also pretty small, so no need to have them
in separate modules.

Block's attribute STRUCTURE has also changed, with the attribute
UNITS added. STRUCTURE now represents every column as a list,
while UNITS represents the 'structure' as
just the objects (e.g. enemies).

This distinction is meant to be used later for find enemies
on the bottom of columns who will have the chance to fire.