Allow enemies to fire bullets
Enemies are now allowed to fire bullets and destroy the player.
This was done because the original game had these features.

Biggest change to the codebase was in a redesign
in how bullets are tracked.
As before the original way of tracking bullets
only considered the player as the only source of bullets.