Change executable to distribute
The executable is different than the one in 'v2.0.0'.

The executable released in v2.0.0 does not work unless
the user performs a git clone.
Hence this is an attempt to resolve this issue.

Using PyInstaller, the runtime-tmpdir
is now pointed to the SpaceInvaders directory.

That, and mixer is now initialized in core.